move ecube gen to utils
